UnitTest project and changed sensors JSON format
Added UnitTests project and ValueUnit value in sensors JSON format.

+++ DataHub/DbCommunicationLib/Model/Sensor.cs
@@ -1,5 +1,7 @@
 ﻿using System.Collections.Generic;
 
+#nullable disable
+
 namespace DbCommunicationLib.Model
 {
     public partial class Sensor
@@ -13,7 +15,7 @@ namespace DbCommunicationLib.Model
         public string Description { get; set; }
         public bool? IsActive { get; set; }
         public string Name { get; set; }
-        public short Type { get; set; }
+        public int Type { get; set; }
         public bool? InverseOnOffLogic { get; set; }
 
         public virtual ICollection<SensorEvent> SensorEvents { get; set; }
